+++ packages/runtime/src/runtime/menu.tsx
@@ -1,8 +1,9 @@
-import type { ReactNode } from "react";
-import React, { useState, useEffect } from "react";
+import type { PropsWithChildren, ReactNode } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import { Menu as AntMenu, Col, Divider } from "antd";
 import * as MuiIcons from "@mui/icons-material";
 import {
+  unwrapWidget,
   useRegisterBindings,
   type EnsembleWidget,
 } from "@ensembleui/react-framework";
@@ -32,9 +33,10 @@ type TypeColors =
 interface MenuItem {
   id?: string;
   testId?: string;
-  icon?: string;
+  icon?: string | { [key: string]: unknown };
+  activeIcon?: string | { [key: string]: unknown };
   iconLibrary?: "default" | "fontAwesome";
-  label: string;
+  label?: string;
   url?: string;
   page?: string;
   selected?: boolean;
@@ -42,6 +44,10 @@ interface MenuItem {
   hasNotifications?: boolean;
   openNewTab?: boolean;
   visible?: boolean;
+  customItem?: {
+    widget?: { [key: string]: unknown };
+    selectedWidget?: { [key: string]: unknown };
+  };
 }
 
 interface MenuBaseProps {
@@ -63,7 +69,6 @@ interface MenuBaseProps {
   };
   header?: EnsembleWidget;
   footer?: EnsembleWidget;
-  enableSearch?: boolean;
 }
 
 const renderMuiIcon = (
@@ -89,13 +94,49 @@ const renderMuiIcon = (
   return null;
 };
 
+const CustomLink: React.FC<PropsWithChildren & { item: MenuItem }> = ({
+  item,
+  children,
+}) => {
+  const content = (
+    <>
+      {children}
+      {item.hasNotifications ? (
+        <div
+          style={{
+            marginTop: "8px",
+            marginLeft: "2px",
+            width: "8px",
+            height: "8px",
+            backgroundColor: "#e07407",
+            borderRadius: "50%",
+          }}
+        />
+      ) : null}
+    </>
+  );
+
+  if (!item.page && !item.url) {
+    return <div>{content}</div>;
+  }
+
+  return (
+    <Link
+      target={item.openNewTab ? "_blank" : "_self"}
+      to={item.page ? `/${item.page}` : String(item.url)}
+    >
+      {content}
+    </Link>
+  );
+};
+
 export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
-  const [isCollapsed, setCollapsed] = useState<boolean>(false);
+  const [isCollapsed, setIsCollapsed] = useState<boolean>(false);
   const { values } = useRegisterBindings({ ...props, isCollapsed }, id, {
-    setCollapsed,
+    setCollapsed: setIsCollapsed,
   });
   const location = useLocation();
-  const [selectedItem, setSelectedItem] = useState<string | null>(null);
+  const [selectedItem, setSelectedItem] = useState<string | undefined>();
 
   const backgroundColor = props.styles?.backgroundColor
     ? getColor(props.styles.backgroundColor)
@@ -104,7 +145,7 @@ export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
   useEffect(() => {
     const initiallySelectedItem = props.items.find((item) => item.selected);
     if (initiallySelectedItem) {
-      setSelectedItem(initiallySelectedItem.label);
+      setSelectedItem(initiallySelectedItem.page);
     }
   }, [props.items]);
 
@@ -114,9 +155,55 @@ export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
         item.page && `/${item.page.toLowerCase()}` === location.pathname,
     );
     if (locationMatch) {
-      setSelectedItem(locationMatch.label);
+      setSelectedItem(locationMatch.page);
     }
   }, [location.pathname, props.items]);
+
+  const getIcon = useCallback(
+    (item: MenuItem) => {
+      const key = selectedItem === item.page ? "activeIcon" : "icon";
+      const icon =
+        selectedItem === item.page && item.activeIcon
+          ? item.activeIcon
+          : item.icon;
+
+      if (!icon) {
+        return null;
+      }
+
+      if (typeof icon === "string") {
+        return renderMuiIcon(
+          icon,
+          props.styles?.iconWidth,
+          props.styles?.iconHeight,
+        );
+      }
+      return EnsembleRuntime.render([
+        {
+          ...unwrapWidget({ Icon: icon }),
+          key,
+        },
+      ]);
+    },
+    [props.styles?.iconHeight, props.styles?.iconWidth, selectedItem],
+  );
+
+  const getLabel = useCallback(
+    (item: MenuItem) => {
+      if (isCollapsed) {
+        return null;
+      }
+      if (!item.customItem) {
+        return item.label;
+      }
+      const widget =
+        selectedItem === item.page && item.customItem.selectedWidget
+          ? item.customItem.selectedWidget
+          : item.customItem.widget;
+      return EnsembleRuntime.render([unwrapWidget(widget || {})]);
+    },
+    [isCollapsed, selectedItem],
+  );
 
   return (
     <Col
@@ -133,11 +220,7 @@ export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
         zIndex: 1,
       }}
     >
-      {props.header ? (
-        <Col style={{ width: "100%" }}>
-          {EnsembleRuntime.render([props.header])}
-        </Col>
-      ) : null}
+      {props.header ? EnsembleRuntime.render([props.header]) : null}
       <AntMenu
         mode="inline"
         /* FIXME This is a hack so we can control our own selected styling. Ideally, this should use design tokens */
@@ -150,32 +233,28 @@ export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
         }}
       >
         {/* FIXME: just use props here https://ant.design/components/menu#examples */}
-        {values?.items.map((item) => (
+        {values?.items.map((item, itemIndex) => (
           <>
             <AntMenu.Item
               data-testid={item.id ?? item.testId}
-              icon={renderMuiIcon(
-                item.icon,
-                props.styles?.iconWidth,
-                props.styles?.iconHeight,
-              )}
-              key={item.page}
+              icon={getIcon(item)}
+              key={item.page || item.url || `customItem${itemIndex}`}
               onClick={(): void => {
-                if (!item.openNewTab) {
-                  setSelectedItem(item.label);
+                if (!item.openNewTab && item.page) {
+                  setSelectedItem(item.page);
                 }
               }}
               style={{
                 color:
-                  selectedItem === item.label
-                    ? (values?.styles?.selectedColor as string) ?? "white"
-                    : (values?.styles?.labelColor as string) ?? "grey",
+                  selectedItem === item.page
+                    ? (values.styles?.selectedColor as string) ?? "white"
+                    : (values.styles?.labelColor as string) ?? "grey",
                 display: item.visible === false ? "none" : "flex",
                 justifyContent: "center",
                 borderRadius: 0,
                 alignItems: "center",
                 fontSize:
-                  selectedItem === item.label
+                  selectedItem === item.page
                     ? `${
                         parseInt(
                           `${
@@ -190,29 +269,12 @@ export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
                           ? props.styles.labelFontSize
                           : 1
                       }rem`,
-                ...(selectedItem === item.label
+                ...(selectedItem === item.page
                   ? props.styles?.onSelectStyles ?? {}
                   : {}),
               }}
             >
-              <Link
-                target={item.openNewTab ? "_blank" : "_self"}
-                to={item.page ? `/${item.page}` : String(item.url)}
-              >
-                {!isCollapsed && item.label}
-                {item.hasNotifications ? (
-                  <div
-                    style={{
-                      marginTop: "8px",
-                      marginLeft: "2px",
-                      width: "8px",
-                      height: "8px",
-                      backgroundColor: "#e07407",
-                      borderRadius: "50%",
-                    }}
-                  />
-                ) : null}
-              </Link>
+              <CustomLink item={item}>{getLabel(item)}</CustomLink>
             </AntMenu.Item>
             {item.divider ? (
               <Col
@@ -231,11 +293,7 @@ export const SideBarMenu: React.FC<MenuBaseProps> = ({ id, ...props }) => {
           </>
         ))}
       </AntMenu>
-      {props.footer ? (
-        <Col style={{ width: "100%" }}>
-          {EnsembleRuntime.render([props.footer])}
-        </Col>
-      ) : null}
+      {props.footer ? EnsembleRuntime.render([props.footer]) : null}
     </Col>
   );
 };