+++ packages/runtime/src/widgets/Form/Dropdown.tsx
@@ -49,6 +49,8 @@ export interface SelectOption {
 
 export type DropdownProps = {
   items?: SelectOption[];
+  /* deprecated, use onChange */
+  onItemSelect: EnsembleAction;
   onChange?: EnsembleAction;
   autoComplete: Expression<boolean>;
   hintStyle?: EnsembleWidgetStyles;
@@ -62,6 +64,7 @@ const Dropdown: React.FC<DropdownProps> = (props) => {
   const [selectedValue, setSelectedValue] = useState<
     string | number | undefined
   >();
+
   const [dropdownStat, setDropdownStat] = useState<boolean>(false);
   const { "item-template": itemTemplate, ...rest } = props;
 
@@ -78,13 +81,16 @@ const Dropdown: React.FC<DropdownProps> = (props) => {
     },
   );
 
-  const action = useEnsembleAction(props.onChange);
-  const handleChange = useCallback(
+  const onItemSelectAction = useEnsembleAction(onItemSelect);
+  const onChangeAction = useEnsembleAction(onChange);
+
+  const onSelectCallback = useCallback(
     (value?: number | string) => {
       setSelectedValue(value);
-      action?.callback({ value });
+      onItemSelectAction?.callback({ selectedValue: value });
+      onChangeAction?.callback({ value });
     },
-    [action],
+    [onItemSelectAction],
   );
 
   const { namedData } = useTemplateData({
@@ -104,7 +110,10 @@ const Dropdown: React.FC<DropdownProps> = (props) => {
               label={isString(item.label) ? item.label : ""}
             >
               {item.items?.map((subItem) => (
-                <Select.Option key={subItem.value}>
+                <Select.Option
+                  key={subItem.value}
+                  onClick={() => onSelectCallback(subItem.value)}
+                >
                   {isString(subItem.label)
                     ? subItem.label
                     : EnsembleRuntime.render([unwrapWidget(subItem.label)])}